Split XPath functions into grouped source files (#531)
