+++ src/xml/xpath/xpath_functions.cpp
@@ -1897,2101 +1897,10 @@ size_t XPathFunctionLibrary::estimate_translate_size(const std::string &Source, 
 //********************************************************************************************************************
 // Core XPath 1.0 Function Implementations
 
-XPathValue XPathFunctionLibrary::function_last(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   return XPathValue((double)Context.size);
-}
-
-XPathValue XPathFunctionLibrary::function_position(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   return XPathValue((double)Context.position);
-}
-
-XPathValue XPathFunctionLibrary::function_count(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   if (Args.size() != 1) return XPathValue(0.0);
-   if (Args[0].type != XPathValueType::NodeSet) return XPathValue(0.0);
-   return XPathValue((double)Args[0].node_set.size());
-}
-
-XPathValue XPathFunctionLibrary::function_id(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   std::vector<XMLTag *> results;
-
-   if (Args.empty()) return XPathValue(results);
-
-   std::unordered_set<std::string> requested_ids;
-
-   auto add_tokens = [&](const std::string &Value) {
-      size_t start = Value.find_first_not_of(" \t\r\n");
-      while (start != std::string::npos) {
-         size_t end = Value.find_first_of(" \t\r\n", start);
-         std::string token = Value.substr(start, (end IS std::string::npos) ? std::string::npos : end - start);
-         if (not token.empty()) requested_ids.insert(token);
-         if (end IS std::string::npos) break;
-         start = Value.find_first_not_of(" \t\r\n", end);
-      }
-   };
-
-   auto collect_from_value = [&](const XPathValue &Value) {
-      switch (Value.type) {
-         case XPathValueType::NodeSet: {
-            if (not Value.node_set_string_values.empty()) {
-               for (const auto &entry : Value.node_set_string_values) add_tokens(entry);
-            }
-            else if (Value.node_set_string_override.has_value()) add_tokens(*Value.node_set_string_override);
-            else {
-               for (auto *node : Value.node_set) {
-                  if (not node) continue;
-                  add_tokens(node->getContent());
-               }
-            }
-            break;
-         }
-
-         case XPathValueType::String:
-         case XPathValueType::Date:
-         case XPathValueType::Time:
-         case XPathValueType::DateTime:
-            add_tokens(Value.string_value);
-            break;
-
-         case XPathValueType::Boolean:
-            add_tokens(Value.to_string());
-            break;
-
-         case XPathValueType::Number:
-            if (not std::isnan(Value.number_value)) add_tokens(Value.to_string());
-            break;
-      }
-   };
-
-   for (const auto &arg : Args) collect_from_value(arg);
-
-   if (requested_ids.empty()) return XPathValue(results);
-   if (not Context.document) return XPathValue(results);
-
-   std::unordered_set<int> seen_tags;
-
-   std::function<void(XMLTag &)> visit = [&](XMLTag &Tag) {
-      if (Tag.isTag()) {
-         for (size_t index = 1; index < Tag.Attribs.size(); ++index) {
-            const auto &attrib = Tag.Attribs[index];
-            if (not (pf::iequals(attrib.Name, "id") or pf::iequals(attrib.Name, "xml:id"))) continue;
-
-            size_t start = attrib.Value.find_first_not_of(" \t\r\n");
-            while (start != std::string::npos) {
-               size_t end = attrib.Value.find_first_of(" \t\r\n", start);
-               std::string token = attrib.Value.substr(start, (end IS std::string::npos) ? std::string::npos : end - start);
-               if (not token.empty() and (requested_ids.find(token) != requested_ids.end())) {
-                  if (seen_tags.insert(Tag.ID).second) results.push_back(&Tag);
-                  break;
-               }
-               if (end IS std::string::npos) break;
-               start = attrib.Value.find_first_not_of(" \t\r\n", end);
-            }
-         }
-      }
-
-      for (auto &child : Tag.Children) visit(child);
-   };
-
-   for (auto &root : Context.document->Tags) visit(root);
-
-   return XPathValue(results);
-}
-
-XPathValue XPathFunctionLibrary::function_local_name(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   XMLTag *target_node = nullptr;
-   const XMLAttrib *target_attribute = nullptr;
-
-   if (Args.empty()) {
-      target_node = Context.context_node;
-      target_attribute = Context.attribute_node;
-   }
-   else if (Args[0].type IS XPathValueType::NodeSet) target_node = Args[0].node_set.empty() ? nullptr : Args[0].node_set[0];
-   else return XPathValue(std::string());
-
-   if (target_attribute) {
-      std::string_view name = target_attribute->Name;
-      auto colon = name.find(':');
-      if (colon IS std::string::npos) return XPathValue(std::string(name));
-      return XPathValue(std::string(name.substr(colon + 1)));
-   }
-
-   if (not target_node) return XPathValue(std::string());
-   if (target_node->Attribs.empty()) return XPathValue(std::string());
-
-   std::string_view node_name = target_node->Attribs[0].Name;
-   if (node_name.empty()) return XPathValue(std::string());
-
-   auto colon = node_name.find(':');
-   if (colon IS std::string::npos) return XPathValue(std::string(node_name));
-   return XPathValue(std::string(node_name.substr(colon + 1)));
-}
-
-XPathValue XPathFunctionLibrary::function_namespace_uri(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   XMLTag *target_node = nullptr;
-   const XMLAttrib *target_attribute = nullptr;
-
-   if (Args.empty()) {
-      target_node = Context.context_node;
-      target_attribute = Context.attribute_node;
-   }
-   else if (Args[0].type IS XPathValueType::NodeSet) target_node = Args[0].node_set.empty() ? nullptr : Args[0].node_set[0];
-   else return XPathValue(std::string());
-
-   if (target_attribute) {
-      std::string_view name = target_attribute->Name;
-      auto colon = name.find(':');
-      if (colon IS std::string::npos) return XPathValue(std::string());
-
-      std::string prefix(name.substr(0, colon));
-      if (pf::iequals(prefix, "xml")) return XPathValue("http://www.w3.org/XML/1998/namespace");
-      if (pf::iequals(prefix, "xmlns")) return XPathValue("http://www.w3.org/2000/xmlns/");
-
-      XMLTag *scope_node = target_node ? target_node : Context.context_node;
-      if (not scope_node) return XPathValue(std::string());
-
-      if (Context.document) {
-         std::string uri = find_in_scope_namespace(scope_node, Context.document, prefix);
-         return XPathValue(uri);
-      }
-
-      return XPathValue(std::string());
-   }
-
-   if (not target_node) return XPathValue(std::string());
-
-   std::string prefix;
-   if (not target_node->Attribs.empty()) {
-      std::string_view node_name = target_node->Attribs[0].Name;
-      auto colon = node_name.find(':');
-      if (colon != std::string::npos) prefix = std::string(node_name.substr(0, colon));
-   }
-
-   if (not prefix.empty()) {
-      if (pf::iequals(prefix, "xml")) return XPathValue("http://www.w3.org/XML/1998/namespace");
-      if (pf::iequals(prefix, "xmlns")) return XPathValue("http://www.w3.org/2000/xmlns/");
-   }
-
-   if ((target_node->NamespaceID != 0) and Context.document) {
-      if (auto uri = Context.document->getNamespaceURI(target_node->NamespaceID)) return XPathValue(*uri);
-   }
-
-   if (Context.document) {
-      std::string uri;
-      if (not prefix.empty()) uri = find_in_scope_namespace(target_node, Context.document, prefix);
-      else uri = find_in_scope_namespace(target_node, Context.document, std::string());
-      return XPathValue(uri);
-   }
-
-   return XPathValue(std::string());
-}
-
-XPathValue XPathFunctionLibrary::function_name(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   XMLTag *target_node = nullptr;
-   const XMLAttrib *target_attribute = nullptr;
-
-   if (Args.empty()) {
-      target_node = Context.context_node;
-      target_attribute = Context.attribute_node;
-   }
-   else if (Args[0].type IS XPathValueType::NodeSet) target_node = Args[0].node_set.empty() ? nullptr : Args[0].node_set[0];
-   else return XPathValue(std::string());
-
-   if (target_attribute) return XPathValue(target_attribute->Name);
-
-   if (not target_node) return XPathValue(std::string());
-   if (target_node->Attribs.empty()) return XPathValue(std::string());
-
-   return XPathValue(target_node->Attribs[0].Name);
-}
-
-XPathValue XPathFunctionLibrary::function_string(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) {
-      if (Context.attribute_node) {
-         return XPathValue(Context.attribute_node->Value);
-      }
-
-      if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_set_value(nodes);
-         return XPathValue(node_set_value.to_string());
-      }
-      return XPathValue(std::string());
-   }
-   return XPathValue(Args[0].to_string());
-}
-
-XPathValue XPathFunctionLibrary::function_concat(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   // Pre-calculate total length to avoid quadratic behavior
-   size_t total_length = 0;
-   std::vector<std::string> arg_strings;
-   arg_strings.reserve(Args.size());
-
-   for (const auto &arg : Args) {
-      arg_strings.emplace_back(arg.to_string());
-      total_length += arg_strings.back().length();
-   }
-
-   std::string result;
-   result.reserve(total_length);
-
-   for (const auto &str : arg_strings) {
-      result += str;
-   }
-
-   return XPathValue(result);
-}
-
-XPathValue XPathFunctionLibrary::function_codepoints_to_string(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::string());
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-   if (length IS 0u) return XPathValue(std::string());
-
-   std::string output;
-   output.reserve(length * 4u);
-
-   for (size_t index = 0u; index < length; ++index) {
-      XPathValue item = extract_sequence_item(sequence, index);
-      double numeric = item.to_number();
-      if (std::isnan(numeric)) continue;
-
-      long long rounded = (long long)std::llround(numeric);
-      if (rounded < 0) {
-         append_codepoint_utf8(output, 0xFFFDu);
-         continue;
-      }
-
-      append_codepoint_utf8(output, (uint32_t)rounded);
-   }
-
-   (void)Context;
-   return XPathValue(output);
-}
-
-XPathValue XPathFunctionLibrary::function_string_to_codepoints(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::vector<XMLTag *>());
-
-   std::string input = Args[0].to_string();
-
-   SequenceBuilder builder;
-   size_t offset = 0u;
-
-   while (offset < input.length()) {
-      int length = 0;
-      uint32_t code = UTF8ReadValue(input.c_str() + offset, &length);
-      if (length <= 0) {
-         code = (unsigned char)input[offset];
-         length = 1;
-      }
-
-      builder.nodes.push_back(nullptr);
-      builder.attributes.push_back(nullptr);
-      builder.strings.push_back(std::to_string(code));
-
-      offset += (size_t)length;
-   }
-
-   (void)Context;
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_compare(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue();
-   if (Args[0].is_empty() or Args[1].is_empty()) return XPathValue();
-
-   std::string left = Args[0].to_string();
-   std::string right = Args[1].to_string();
-   std::string collation = (Args.size() > 2u) ? Args[2].to_string() : std::string();
-
-   if (!collation.empty() and (collation != "http://www.w3.org/2005/xpath-functions/collation/codepoint") and
-       (collation != "unicode")) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue();
-   }
-
-   int result = 0;
-   if (left < right) result = -1;
-   else if (left IS right) result = 0;
-   else result = 1;
-
-   return XPathValue((double)result);
-}
-
-XPathValue XPathFunctionLibrary::function_codepoint_equal(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue();
-   if (Args[0].is_empty() or Args[1].is_empty()) return XPathValue();
-
-   std::string first = Args[0].to_string();
-   std::string second = Args[1].to_string();
-
-   (void)Context;
-   return XPathValue(first IS second);
-}
-
-XPathValue XPathFunctionLibrary::function_starts_with(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 2) return XPathValue(false);
-   std::string str = Args[0].to_string();
-   std::string prefix = Args[1].to_string();
-   return XPathValue(str.substr(0, prefix.length()) IS prefix);
-}
-
-XPathValue XPathFunctionLibrary::function_ends_with(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 2u) return XPathValue(false);
-
-   std::string input = Args[0].to_string();
-   std::string suffix = Args[1].to_string();
-   if (suffix.length() > input.length()) return XPathValue(false);
-
-   return XPathValue(input.compare(input.length() - suffix.length(), suffix.length(), suffix) IS 0);
-}
-
-XPathValue XPathFunctionLibrary::function_contains(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 2) return XPathValue(false);
-   std::string str = Args[0].to_string();
-   std::string substr = Args[1].to_string();
-   return XPathValue(str.find(substr) != std::string::npos);
-}
-
-XPathValue XPathFunctionLibrary::function_substring_before(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 2) return XPathValue(std::string());
-
-   std::string source = Args[0].to_string();
-   std::string pattern = Args[1].to_string();
-
-   if (pattern.empty()) return XPathValue(std::string());
-
-   size_t position = source.find(pattern);
-   if (position IS std::string::npos) return XPathValue(std::string());
-
-   return XPathValue(source.substr(0, position));
-}
-
-XPathValue XPathFunctionLibrary::function_substring_after(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 2) return XPathValue(std::string());
-
-   std::string source = Args[0].to_string();
-   std::string pattern = Args[1].to_string();
-
-   if (pattern.empty()) return XPathValue(source);
-
-   size_t position = source.find(pattern);
-   if (position IS std::string::npos) return XPathValue(std::string());
-
-   return XPathValue(source.substr(position + pattern.length()));
-}
-
-XPathValue XPathFunctionLibrary::function_substring(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2 or Args.size() > 3) return XPathValue(std::string());
-
-   std::string str = Args[0].to_string();
-   if (str.empty()) return XPathValue(std::string());
-
-   double start_pos = Args[1].to_number();
-   if (std::isnan(start_pos) or std::isinf(start_pos)) return XPathValue(std::string());
-
-   // XPath uses 1-based indexing
-   int start_index = (int)std::round(start_pos) - 1;
-   if (start_index < 0) start_index = 0;
-   if (start_index >= (int)str.length()) return XPathValue(std::string());
-
-   if (Args.size() IS 3) {
-      double length = Args[2].to_number();
-      if (std::isnan(length) or std::isinf(length) or length <= 0) return XPathValue(std::string());
-
-      int len = (int)std::round(length);
-      int remaining = (int)str.length() - start_index;
-      if (len > remaining) len = remaining;
-
-      // For small substrings, avoid extra allocation overhead
-      if (len <= 0) return XPathValue(std::string());
-
-      return XPathValue(str.substr(start_index, len));
-   }
-
-   // Return substring from start_index to end
-   return XPathValue(str.substr(start_index));
-}
-
-XPathValue XPathFunctionLibrary::function_string_length(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string str;
-   if (Args.empty()) {
-      if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_set_value(nodes);
-         str = node_set_value.to_string();
-      }
-   }
-   else str = Args[0].to_string();
-   return XPathValue((double)str.length());
-}
-
-XPathValue XPathFunctionLibrary::function_normalize_space(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string str;
-   if (Args.empty()) {
-      if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_set_value(nodes);
-         str = node_set_value.to_string();
-      }
-   }
-   else str = Args[0].to_string();
-
-   // Remove leading and trailing whitespace, collapse internal whitespace
-   size_t start = str.find_first_not_of(" \t\n\r");
-   if (start IS std::string::npos) return XPathValue(std::string());
-
-   size_t end = str.find_last_not_of(" \t\n\r");
-   str = str.substr(start, end - start + 1);
-
-   // Collapse internal whitespace
-   std::string result;
-   result.reserve(estimate_normalize_space_size(str));
-   bool in_whitespace = false;
-   for (char c : str) {
-      if (c IS ' ' or c IS '\t' or c IS '\n' or c IS '\r') {
-         if (not in_whitespace) {
-            result += ' ';
-            in_whitespace = true;
-         }
-      }
-      else {
-         result += c;
-         in_whitespace = false;
-      }
-   }
-
-   return XPathValue(result);
-}
-
-XPathValue XPathFunctionLibrary::function_normalize_unicode(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::string());
-
-   std::string input = Args[0].to_string();
-   std::string form = (Args.size() > 1u) ? Args[1].to_string() : std::string("NFC");
-
-   bool unsupported = false;
-   std::string normalised = simple_normalise_unicode(input, form, &unsupported);
-   if (unsupported and Context.expression_unsupported) *Context.expression_unsupported = true;
-
-   return XPathValue(normalised);
-}
-
-XPathValue XPathFunctionLibrary::function_string_join(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::string());
-
-   const XPathValue &sequence = Args[0];
-   std::string separator = (Args.size() > 1u) ? Args[1].to_string() : std::string();
-
-   size_t length = sequence_length(sequence);
-   if (length IS 0u) return XPathValue(std::string());
-
-   std::string result;
-   for (size_t index = 0u; index < length; ++index) {
-      if (index > 0u) result.append(separator);
-      result.append(sequence_item_string(sequence, index));
-   }
-
-   (void)Context;
-   return XPathValue(result);
-}
-
-XPathValue XPathFunctionLibrary::function_translate(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   if (Args.size() != 3) return XPathValue(std::string());
-
-   std::string source = Args[0].to_string();
-   std::string from = Args[1].to_string();
-   std::string to = Args[2].to_string();
-
-   if (source.empty()) return XPathValue(std::string());
-
-   // Pre-size result string based on input length (worst case: no characters removed)
-   std::string result;
-   result.reserve(source.size());
-
-   // Build character mapping for faster lookups if worth it
-   if (from.size() > 10) {
-      // Use array mapping for better performance with larger translation sets
-      std::array<int, 256> char_map;
-      char_map.fill(-1);
-
-      for (size_t i = 0; i < from.size() and i < 256; ++i) {
-         unsigned char ch = (unsigned char)from[i];
-         if (char_map[ch] IS -1) { // First occurrence takes precedence
-            char_map[ch] = (int)i;
-         }
-      }
-
-      for (char ch : source) {
-         unsigned char uch = (unsigned char)ch;
-         int index = char_map[uch];
-         if (index IS -1) result.push_back(ch);
-         else if (index < (int)to.length()) result.push_back(to[index]);
-      }
-   }
-   else {
-      // Use simple find for small translation sets
-      for (char ch : source) {
-         size_t index = from.find(ch);
-         if (index IS std::string::npos) result.push_back(ch);
-         else if (index < to.length()) result.push_back(to[index]);
-      }
-   }
-
-   return XPathValue(result);
-}
-
-XPathValue XPathFunctionLibrary::function_upper_case(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string input;
-
-   if (Args.empty()) {
-      if (Context.attribute_node) input = Context.attribute_node->Value;
-      else if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_value(nodes);
-         input = node_value.to_string();
-      }
-      else input = std::string();
-   }
-   else input = Args[0].to_string();
-
-   return XPathValue(apply_string_case(input, true));
-}
-
-XPathValue XPathFunctionLibrary::function_lower_case(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string input;
-
-   if (Args.empty()) {
-      if (Context.attribute_node) input = Context.attribute_node->Value;
-      else if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_value(nodes);
-         input = node_value.to_string();
-      }
-      else input = std::string();
-   }
-   else input = Args[0].to_string();
-
-   return XPathValue(apply_string_case(input, false));
-}
-
-XPathValue XPathFunctionLibrary::function_iri_to_uri(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string input;
-
-   if (Args.empty()) {
-      if (Context.attribute_node) input = Context.attribute_node->Value;
-      else if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_value(nodes);
-         input = node_value.to_string();
-      }
-      else input = std::string();
-   }
-   else input = Args[0].to_string();
-
-   std::string result;
-   result.reserve(input.length() * 3u);
-
-   static const char hex_digits[] = "0123456789ABCDEF";
-
-   for (unsigned char code : input) {
-      if (code <= 0x7Fu) {
-         result.push_back((char)code);
-      }
-      else {
-         result.push_back('%');
-         result.push_back(hex_digits[(code >> 4u) & 0x0Fu]);
-         result.push_back(hex_digits[code & 0x0Fu]);
-      }
-   }
-
-   return XPathValue(result);
-}
-
-XPathValue XPathFunctionLibrary::function_encode_for_uri(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string input;
-
-   if (Args.empty()) {
-      if (Context.attribute_node) input = Context.attribute_node->Value;
-      else if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_value(nodes);
-         input = node_value.to_string();
-      }
-      else input = std::string();
-   }
-   else input = Args[0].to_string();
-
-   return XPathValue(encode_for_uri_impl(input));
-}
-
-XPathValue XPathFunctionLibrary::function_escape_html_uri(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string input;
-
-   if (Args.empty()) {
-      if (Context.attribute_node) input = Context.attribute_node->Value;
-      else if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_value(nodes);
-         input = node_value.to_string();
-      }
-      else input = std::string();
-   }
-   else input = Args[0].to_string();
-
-   return XPathValue(escape_html_uri_impl(input));
-}
-
-XPathValue XPathFunctionLibrary::function_error(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   std::string error_code = "err:FOER0000";
-   std::string description = "User-defined error";
-   std::string detail;
-
-   if (not Args.empty()) {
-      const XPathValue &code_value = Args[0];
-      if (not code_value.is_empty()) error_code = code_value.to_string();
-   }
-
-   if (Args.size() > 1) {
-      const XPathValue &description_value = Args[1];
-      if (not description_value.is_empty()) description = description_value.to_string();
-   }
-
-   if (Args.size() > 2) {
-      const XPathValue &detail_value = Args[2];
-      if (not detail_value.is_empty()) detail = describe_xpath_value(detail_value);
-   }
-
-   pf::Log log(__FUNCTION__);
-
-   if (detail.empty()) {
-      log.warning("XPath error (%s): %s", error_code.c_str(), description.c_str());
-   }
-   else {
-      log.warning("XPath error (%s): %s [%s]", error_code.c_str(), description.c_str(), detail.c_str());
-   }
-
-   if (Context.expression_unsupported) {
-      *Context.expression_unsupported = true;
-   }
-
-   if (Context.document) {
-      if (not Context.document->ErrorMsg.empty()) Context.document->ErrorMsg.append("\n");
-      Context.document->ErrorMsg.append("XPath error ");
-      Context.document->ErrorMsg.append(error_code);
-      Context.document->ErrorMsg.append(": ");
-      Context.document->ErrorMsg.append(description);
-      if (not detail.empty()) {
-         Context.document->ErrorMsg.append(" [");
-         Context.document->ErrorMsg.append(detail);
-         Context.document->ErrorMsg.append("]");
-      }
-   }
-
-   return XPathValue();
-}
-
-XPathValue XPathFunctionLibrary::function_trace(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue();
-
-   const XPathValue &value = Args[0];
-   std::string label = "trace";
-
-   if (Args.size() > 1) {
-      const XPathValue &label_value = Args[1];
-      if (not label_value.is_empty()) label = label_value.to_string();
-   }
-
-   if (label.empty()) label = "trace";
-
-   std::string summary = describe_xpath_value(value);
-   if (summary.empty()) summary = std::string("()");
-
-   pf::Log log(__FUNCTION__);
-   log.msg("XPath trace [%s]: %s", label.c_str(), summary.c_str());
-
-   (void)Context;
-
-   return value;
-}
-
-XPathValue XPathFunctionLibrary::function_matches(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2 or Args.size() > 3) return XPathValue(false);
-
-   std::string input = Args[0].to_string();
-   std::string pattern = Args[1].to_string();
-   std::string flags = (Args.size() IS 3) ? Args[2].to_string() : std::string();
-
-   pf::Regex compiled;
-   if (not compiled.compile(pattern, build_regex_options(flags, Context.expression_unsupported))) {
-      return XPathValue(false);
-   }
-
-   pf::MatchResult result;
-   bool matched = compiled.search(input, result);
-   return XPathValue(matched);
-}
-
-XPathValue XPathFunctionLibrary::function_replace(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 3 or Args.size() > 4) return XPathValue(std::string());
-
-   std::string input = Args[0].to_string();
-   std::string pattern = Args[1].to_string();
-   std::string replacement = Args[2].to_string();
-   std::string flags = (Args.size() IS 4) ? Args[3].to_string() : std::string();
-
-   pf::Regex compiled;
-   if (not compiled.compile(pattern, build_regex_options(flags, Context.expression_unsupported))) {
-      return XPathValue(input);
-   }
-
-   std::string replaced;
-   if (not compiled.replace(input, replacement, replaced)) replaced = input;
-
-   return XPathValue(replaced);
-}
-
-XPathValue XPathFunctionLibrary::function_tokenize(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2 or Args.size() > 3) return XPathValue(std::vector<XMLTag *>());
-
-   std::string input = Args[0].to_string();
-   std::string pattern = Args[1].to_string();
-   std::string flags = (Args.size() IS 3) ? Args[2].to_string() : std::string();
-
-   std::vector<std::string> tokens;
-
-   if (pattern.empty()) {
-      for (size_t index = 0; index < input.length(); ++index) {
-         tokens.emplace_back(input.substr(index, 1));
-      }
-   }
-   else {
-      pf::SyntaxOptions options = build_regex_options(flags, Context.expression_unsupported);
-
-      pf::Regex compiled;
-      if (not compiled.compile(pattern, options)) {
-         return XPathValue(std::vector<XMLTag *>());
-      }
-
-      compiled.tokenize(input, -1, tokens);
-
-      if (not tokens.empty() and tokens.back().empty()) tokens.pop_back();
-   }
-
-   std::vector<XMLTag *> placeholders(tokens.size(), nullptr);
-   return XPathValue(placeholders, std::nullopt, tokens);
-}
-
-XPathValue XPathFunctionLibrary::function_analyze_string(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u or Args.size() > 3u) return XPathValue(std::vector<XMLTag *>());
-
-   std::string input = Args[0].to_string();
-   std::string pattern = Args[1].to_string();
-   std::string flags = (Args.size() > 2u) ? Args[2].to_string() : std::string();
-
-   pf::Regex compiled;
-   if (not compiled.compile(pattern, build_regex_options(flags, Context.expression_unsupported))) {
-      return XPathValue(std::vector<XMLTag *>());
-   }
-
-   SequenceBuilder builder;
-   size_t search_offset = 0u;
-   size_t guard = 0u;
-
-   while (search_offset <= input.length()) {
-      std::string_view remaining(input.c_str() + search_offset, input.length() - search_offset);
-      pf::MatchResult match;
-      if (!compiled.search(remaining, match)) {
-         if (!remaining.empty()) {
-            builder.nodes.push_back(nullptr);
-            builder.attributes.push_back(nullptr);
-            builder.strings.push_back(std::string("non-match:") + std::string(remaining));
-         }
-         break;
-      }
-
-      if ((match.span.offset != (size_t)std::string::npos) and (match.span.offset > 0u)) {
-         std::string unmatched = std::string(remaining.substr(0u, match.span.offset));
-         if (!unmatched.empty()) {
-            builder.nodes.push_back(nullptr);
-            builder.attributes.push_back(nullptr);
-            builder.strings.push_back(std::string("non-match:") + unmatched);
-         }
-      }
-
-      std::string matched_text;
-      if (match.span.offset != (size_t)std::string::npos) {
-         matched_text = std::string(remaining.substr(match.span.offset, match.span.length));
-      }
-
-      builder.nodes.push_back(nullptr);
-      builder.attributes.push_back(nullptr);
-      builder.strings.push_back(std::string("match:") + matched_text);
-
-      for (size_t index = 1u; index < match.captures.size(); ++index) {
-         if (match.capture_spans[index].offset IS (size_t)std::string::npos) continue;
-         builder.nodes.push_back(nullptr);
-         builder.attributes.push_back(nullptr);
-         builder.strings.push_back(std::string("group") + std::to_string(index) + std::string(":") + match.captures[index]);
-      }
-
-      size_t advance = 0u;
-      if (match.span.offset != (size_t)std::string::npos) advance = match.span.offset;
-      if (match.span.length > 0u) advance += match.span.length;
-      else advance += 1u;
-
-      search_offset += advance;
-
-      guard++;
-      if (guard > input.length() + 8u) break;
-   }
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_resolve_uri(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue();
-
-   std::string relative = Args[0].to_string();
-   std::string base;
-
-   if (Args.size() > 1u and not Args[1].is_empty()) base = Args[1].to_string();
-   else if (Context.document) base = Context.document->Path;
-
-   if (relative.empty()) {
-      if (base.empty()) return XPathValue();
-      return XPathValue(base);
-   }
-
-   if (is_absolute_uri(relative)) return XPathValue(relative);
-   if (base.empty()) return XPathValue();
-
-   std::string resolved = resolve_relative_uri(relative, base);
-   return XPathValue(resolved);
-}
-
-XPathValue XPathFunctionLibrary::function_boolean(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(false);
-   return XPathValue(Args[0].to_boolean());
-}
-
-XPathValue XPathFunctionLibrary::function_not(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(true);
-   return XPathValue(!Args[0].to_boolean());
-}
-
-XPathValue XPathFunctionLibrary::function_true(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   return XPathValue(true);
-}
-
-XPathValue XPathFunctionLibrary::function_false(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   return XPathValue(false);
-}
-
-XPathValue XPathFunctionLibrary::function_lang(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(false);
-
-   std::string requested = Args[0].to_string();
-   if (requested.empty()) return XPathValue(false);
-
-   XMLTag *node = Context.context_node;
-   if (not node) return XPathValue(false);
-
-   std::string language = find_language_for_node(node, Context.document);
-   if (language.empty()) return XPathValue(false);
-
-   return XPathValue(language_matches(language, requested));
-}
-
-XPathValue XPathFunctionLibrary::function_exists(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(false);
-
-   const XPathValue &value = Args[0];
-
-   if (value.type IS XPathValueType::NodeSet) {
-      if (not value.node_set.empty()) return XPathValue(true);
-
-      if (value.node_set_string_override.has_value()) return XPathValue(true);
-
-      if (not value.node_set_string_values.empty()) return XPathValue(true);
-
-      if (not value.node_set_attributes.empty()) return XPathValue(true);
-
-      return XPathValue(false);
-   }
-
-   return XPathValue(true);
-}
-
-XPathValue XPathFunctionLibrary::function_index_of(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2) return XPathValue(std::vector<XMLTag *>());
-
-   if ((Args.size() > 2) and Context.expression_unsupported) *Context.expression_unsupported = true;
-
-   const XPathValue &sequence = Args[0];
-   const XPathValue &lookup = Args[1];
-
-   size_t length = sequence_length(sequence);
-   if (length IS 0) return XPathValue(std::vector<XMLTag *>());
-
-   XPathValue target = extract_sequence_item(lookup, 0);
-   SequenceBuilder builder;
-
-   for (size_t index = 0; index < length; ++index) {
-      XPathValue item = extract_sequence_item(sequence, index);
-      if (xpath_values_equal(item, target)) {
-         builder.nodes.push_back(nullptr);
-         builder.attributes.push_back(nullptr);
-         builder.strings.push_back(format_xpath_number(double(index + 1)));
-      }
-   }
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_empty(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(true);
-
-   size_t length = sequence_length(Args[0]);
-   return XPathValue(length IS 0);
-}
-
-XPathValue XPathFunctionLibrary::function_distinct_values(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::vector<XMLTag *>());
-
-   if ((Args.size() > 1) and Context.expression_unsupported) *Context.expression_unsupported = true;
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-   if (length IS 0) return XPathValue(std::vector<XMLTag *>());
-
-   std::unordered_set<std::string> seen;
-   SequenceBuilder builder;
-
-   for (size_t index = 0; index < length; ++index) {
-      std::string key = sequence_item_string(sequence, index);
-      auto insert_result = seen.insert(key);
-      if (insert_result.second) {
-         XPathValue item = extract_sequence_item(sequence, index);
-         append_value_to_sequence(item, builder);
-      }
-   }
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_insert_before(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 3) {
-      if (Args.empty()) return XPathValue(std::vector<XMLTag *>());
-      return Args[0];
-   }
-
-   const XPathValue &sequence = Args[0];
-   double position_value = Args[1].to_number();
-   const XPathValue &insertion = Args[2];
-
-   size_t length = sequence_length(sequence);
-   size_t insert_index = 0;
-
-   if (std::isnan(position_value)) insert_index = 0;
-   else if (std::isinf(position_value)) insert_index = (position_value > 0.0) ? length : 0;
-   else {
-      long long floored = (long long)std::floor(position_value);
-      if (floored <= 1) insert_index = 0;
-      else if (floored > (long long)length) insert_index = length;
-      else insert_index = (size_t)(floored - 1);
-   }
-
-   if (insert_index > length) insert_index = length;
-
-   SequenceBuilder builder;
-
-   for (size_t index = 0; index < length; ++index) {
-      if (index IS insert_index) append_value_to_sequence(insertion, builder);
-      XPathValue item = extract_sequence_item(sequence, index);
-      append_value_to_sequence(item, builder);
-   }
-
-   if (insert_index >= length) append_value_to_sequence(insertion, builder);
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_remove(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2) {
-      if (Args.empty()) return XPathValue(std::vector<XMLTag *>());
-      return Args[0];
-   }
-
-   const XPathValue &sequence = Args[0];
-   double position_value = Args[1].to_number();
-   size_t length = sequence_length(sequence);
-
-   if (length IS 0) return XPathValue(std::vector<XMLTag *>());
-   if (std::isnan(position_value) or std::isinf(position_value)) return sequence;
-
-   long long floored = (long long)std::floor(position_value);
-   if (floored < 1) return sequence;
-   if (floored > (long long)length) return sequence;
-
-   size_t remove_index = (size_t)(floored - 1);
-   SequenceBuilder builder;
-
-   for (size_t index = 0; index < length; ++index) {
-      if (index IS remove_index) continue;
-      XPathValue item = extract_sequence_item(sequence, index);
-      append_value_to_sequence(item, builder);
-   }
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_reverse(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::vector<XMLTag *>());
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-   SequenceBuilder builder;
-
-   for (size_t index = length; index > 0; --index) {
-      XPathValue item = extract_sequence_item(sequence, index - 1);
-      append_value_to_sequence(item, builder);
-   }
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_subsequence(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2) return XPathValue(std::vector<XMLTag *>());
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-   if (length IS 0) return XPathValue(std::vector<XMLTag *>());
-
-   double start_value = Args[1].to_number();
-   if (std::isnan(start_value)) return XPathValue(std::vector<XMLTag *>());
-
-   double min_position = std::ceil(start_value);
-   if (std::isnan(min_position)) return XPathValue(std::vector<XMLTag *>());
-   if (min_position < 1.0) min_position = 1.0;
-
-   double max_position = std::numeric_limits<double>::infinity();
-   if (Args.size() > 2) {
-      double length_value = Args[2].to_number();
-      if (std::isnan(length_value)) return XPathValue(std::vector<XMLTag *>());
-      if (length_value <= 0.0) return XPathValue(std::vector<XMLTag *>());
-
-      max_position = std::ceil(start_value + length_value);
-      if (std::isnan(max_position)) return XPathValue(std::vector<XMLTag *>());
-   }
-
-   SequenceBuilder builder;
-
-   for (size_t index = 0; index < length; ++index) {
-      double position = double(index + 1);
-      if (position < min_position) continue;
-      if ((not std::isinf(max_position)) and (position >= max_position)) break;
-      XPathValue item = extract_sequence_item(sequence, index);
-      append_value_to_sequence(item, builder);
-   }
-
-   return make_sequence_value(std::move(builder));
-}
-
-XPathValue XPathFunctionLibrary::function_unordered(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::vector<XMLTag *>());
-   return Args[0];
-}
-
-XPathValue XPathFunctionLibrary::function_deep_equal(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2) return XPathValue(false);
-
-   if ((Args.size() > 2) and Context.expression_unsupported) *Context.expression_unsupported = true;
-
-   const XPathValue &left = Args[0];
-   const XPathValue &right = Args[1];
-
-   size_t left_length = sequence_length(left);
-   size_t right_length = sequence_length(right);
-   if (left_length != right_length) return XPathValue(false);
-
-   for (size_t index = 0; index < left_length; ++index) {
-      XPathValue left_item = extract_sequence_item(left, index);
-      XPathValue right_item = extract_sequence_item(right, index);
-      if (not xpath_values_equal(left_item, right_item)) return XPathValue(false);
-   }
-
-   return XPathValue(true);
-}
-
-XPathValue XPathFunctionLibrary::function_zero_or_one(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue();
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-
-   if (length <= 1) return sequence;
-
-   flag_cardinality_error(Context, "zero-or-one", "argument has more than one item");
-   return XPathValue();
-}
-
-XPathValue XPathFunctionLibrary::function_one_or_more(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue();
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-
-   if (length IS 0) {
-      flag_cardinality_error(Context, "one-or-more", "argument is empty");
-      return XPathValue();
-   }
-
-   return sequence;
-}
-
-XPathValue XPathFunctionLibrary::function_exactly_one(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue();
-
-   const XPathValue &sequence = Args[0];
-   size_t length = sequence_length(sequence);
-
-   if (length IS 1) return sequence;
-
-   if (length IS 0) flag_cardinality_error(Context, "exactly-one", "argument is empty");
-   else flag_cardinality_error(Context, "exactly-one", "argument has more than one item");
-
-   return XPathValue();
-}
-
-XPathValue XPathFunctionLibrary::function_number(const std::vector<XPathValue> &Args, const XPathContext &Context) {
-   if (Args.empty()) {
-      if (Context.context_node) {
-         std::vector<XMLTag *> nodes = { Context.context_node };
-         XPathValue node_set_value(nodes);
-         return XPathValue(node_set_value.to_number());
-      }
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-   return XPathValue(Args[0].to_number());
-}
-
-XPathValue XPathFunctionLibrary::function_sum(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(0.0);
-   if (Args[0].type != XPathValueType::NodeSet) return XPathValue(0.0);
-
-   const auto &nodeset = Args[0];
-   double sum = 0.0;
-
-   // Handle attribute nodes if present
-   if (not nodeset.node_set_attributes.empty()) {
-      for (const XMLAttrib *attr : nodeset.node_set_attributes) {
-         if (attr) {
-            double value = XPathValue::string_to_number(attr->Value);
-            if (not std::isnan(value)) {
-               sum += value;
-            }
-         }
-      }
-   }
-   // Handle string values if present
-   else if (not nodeset.node_set_string_values.empty()) {
-      for (const std::string &str : nodeset.node_set_string_values) {
-         double value = XPathValue::string_to_number(str);
-         if (not std::isnan(value)) {
-            sum += value;
-         }
-      }
-   }
-   // Handle regular element nodes
-   else {
-      for (XMLTag *node : nodeset.node_set) {
-         if (node) {
-            std::string node_content = XPathValue::node_string_value(node);
-            double value = XPathValue::string_to_number(node_content);
-            if (not std::isnan(value)) {
-               sum += value;
-            }
-         }
-      }
-   }
-   return XPathValue(sum);
-}
-
-XPathValue XPathFunctionLibrary::function_floor(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   double value = Args[0].to_number();
-   if (std::isnan(value) or std::isinf(value)) return XPathValue(value);
-   return XPathValue(std::floor(value));
-}
-
-XPathValue XPathFunctionLibrary::function_ceiling(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   double value = Args[0].to_number();
-   if (std::isnan(value) or std::isinf(value)) return XPathValue(value);
-   return XPathValue(std::ceil(value));
-}
-
-XPathValue XPathFunctionLibrary::function_round(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   double value = Args[0].to_number();
-   if (std::isnan(value) or std::isinf(value)) return XPathValue(value);
-   return XPathValue(std::round(value));
-}
-
-XPathValue XPathFunctionLibrary::function_round_half_to_even(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args.size() > 2) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = Args[0].to_number();
-   if (std::isnan(value) or std::isinf(value)) return XPathValue(value);
-
-   int precision = 0;
-   if (Args.size() > 1) precision = (int)Args[1].to_number();
-
-   double scaled = value;
-   double factor = 1.0;
-   bool negative_precision = false;
-
-   if (precision > 0) {
-      factor = std::pow(10.0, precision);
-      if (std::isnan(factor) or std::isinf(factor) or factor IS 0.0) return XPathValue(value);
-      scaled = value * factor;
-      if (std::isnan(scaled) or std::isinf(scaled)) return XPathValue(value);
-   }
-   else if (precision < 0) {
-      negative_precision = true;
-      factor = std::pow(10.0, -precision);
-      if (std::isnan(factor) or std::isinf(factor) or factor IS 0.0) return XPathValue(value);
-      scaled = value / factor;
-   }
-
-   double rounded_scaled = std::nearbyint(scaled);
-
-   if (std::isnan(rounded_scaled) or std::isinf(rounded_scaled)) return XPathValue(rounded_scaled);
-
-   double result = rounded_scaled;
-   if (precision > 0) {
-      result = rounded_scaled / factor;
-   }
-   else if (negative_precision) {
-      result = rounded_scaled * factor;
-   }
-
-   return XPathValue(result);
-}
-
-XPathValue XPathFunctionLibrary::function_abs(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() != 1) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = Args[0].to_number();
-   if (std::isnan(value) or std::isinf(value)) return XPathValue(value);
-
-   return XPathValue(std::fabs(value));
-}
-
-XPathValue XPathFunctionLibrary::function_min(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   std::vector<double> numbers;
-   numbers.reserve(Args.size());
-
-   for (const auto &arg : Args) append_numbers_from_value(arg, numbers);
-
-   if (numbers.empty()) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double minimum = numbers[0];
-   for (size_t index = 1; index < numbers.size(); ++index) {
-      if (numbers[index] < minimum) minimum = numbers[index];
-   }
-
-   return XPathValue(minimum);
-}
-
-XPathValue XPathFunctionLibrary::function_max(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   std::vector<double> numbers;
-   numbers.reserve(Args.size());
-
-   for (const auto &arg : Args) append_numbers_from_value(arg, numbers);
-
-   if (numbers.empty()) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double maximum = numbers[0];
-   for (size_t index = 1; index < numbers.size(); ++index) {
-      if (numbers[index] > maximum) maximum = numbers[index];
-   }
-
-   return XPathValue(maximum);
-}
-
-XPathValue XPathFunctionLibrary::function_avg(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.empty()) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   std::vector<double> numbers;
-   numbers.reserve(Args.size());
-
-   for (const auto &arg : Args) append_numbers_from_value(arg, numbers);
-
-   if (numbers.empty()) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double total = 0.0;
-   for (double value : numbers) total += value;
-
-   return XPathValue(total / (double)numbers.size());
-}
-
-XPathValue XPathFunctionLibrary::function_format_date(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue(std::string());
-   if (Args[0].is_empty()) return XPathValue();
-
-   std::string value = Args[0].to_string();
-   std::string picture = Args[1].to_string();
-
-   if (Args.size() > 2u and not Args[2].is_empty()) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-   }
-
-   DateTimeComponents components;
-   if (!parse_date_value(value, components)) return XPathValue(value);
-
-   std::string formatted = format_with_picture(components, picture);
-   return XPathValue(formatted);
-}
-
-XPathValue XPathFunctionLibrary::function_format_time(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue(std::string());
-   if (Args[0].is_empty()) return XPathValue();
-
-   std::string value = Args[0].to_string();
-   std::string picture = Args[1].to_string();
-
-   if (Args.size() > 2u and not Args[2].is_empty()) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-   }
-
-   DateTimeComponents components;
-   if (!parse_time_value(value, components)) return XPathValue(value);
-
-   std::string formatted = format_with_picture(components, picture);
-   return XPathValue(formatted);
-}
-
-XPathValue XPathFunctionLibrary::function_format_date_time(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue(std::string());
-   if (Args[0].is_empty()) return XPathValue();
-
-   std::string value = Args[0].to_string();
-   std::string picture = Args[1].to_string();
-
-   if (Args.size() > 2u and not Args[2].is_empty()) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-   }
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(value, components)) return XPathValue(value);
-
-   std::string formatted = format_with_picture(components, picture);
-   return XPathValue(formatted);
-}
-
-XPathValue XPathFunctionLibrary::function_format_integer(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue(std::string());
-
-   double number = Args[0].to_number();
-   if (std::isnan(number) or std::isinf(number)) return XPathValue(std::string());
-
-   if (Args.size() > 2u and not Args[2].is_empty()) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-   }
-
-   long long rounded = (long long)std::llround(number);
-   std::string picture = Args[1].to_string();
-   std::string formatted = format_integer_picture(rounded, picture);
-   return XPathValue(formatted);
-}
-
-namespace {
-
-std::chrono::system_clock::time_point current_utc_time_point()
-{
-   auto now = std::chrono::system_clock::now();
-   return std::chrono::time_point_cast<std::chrono::seconds>(now);
-}
-
-std::tm make_utc_tm(std::chrono::system_clock::time_point TimePoint)
-{
-   std::time_t raw_time = std::chrono::system_clock::to_time_t(TimePoint);
-   std::tm utc{};
-#if defined(_WIN32)
-   gmtime_s(&utc, &raw_time);
-#else
-   gmtime_r(&raw_time, &utc);
-#endif
-   return utc;
-}
-
-std::string format_utc_date(const std::tm &Tm)
-{
-   char buffer[32];
-   std::snprintf(buffer, sizeof(buffer), "%04d-%02d-%02d", Tm.tm_year + 1900, Tm.tm_mon + 1, Tm.tm_mday);
-   return std::string(buffer);
-}
-
-std::string format_utc_time(const std::tm &Tm)
-{
-   char buffer[32];
-   std::snprintf(buffer, sizeof(buffer), "%02d:%02d:%02d", Tm.tm_hour, Tm.tm_min, Tm.tm_sec);
-   return std::string(buffer);
-}
-
-} // namespace
-
-XPathValue XPathFunctionLibrary::function_current_date(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   auto now = current_utc_time_point();
-   auto tm = make_utc_tm(now);
-   return XPathValue(XPathValueType::Date, format_utc_date(tm));
-}
-
-XPathValue XPathFunctionLibrary::function_current_time(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   auto now = current_utc_time_point();
-   auto tm = make_utc_tm(now);
-   std::string time = format_utc_time(tm);
-   time.push_back('Z');
-   return XPathValue(XPathValueType::Time, std::move(time));
-}
-
-XPathValue XPathFunctionLibrary::function_current_date_time(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   auto now = current_utc_time_point();
-   auto tm = make_utc_tm(now);
-   std::string date = format_utc_date(tm);
-   std::string time = format_utc_time(tm);
-   std::string combined;
-   combined.reserve(date.length() + time.length() + 2);
-   combined.append(date);
-   combined.push_back('T');
-   combined.append(time);
-   combined.push_back('Z');
-   return XPathValue(XPathValueType::DateTime, std::move(combined));
-}
-
-XPathValue XPathFunctionLibrary::function_date_time(const std::vector<XPathValue> &Args, const XPathContext &Context)
-{
-   if (Args.size() < 2u) return XPathValue();
-   if (Args[0].is_empty() or Args[1].is_empty()) return XPathValue();
-
-   std::string date_value = Args[0].to_string();
-   std::string time_value = Args[1].to_string();
-
-   DateTimeComponents combined;
-   if (!combine_date_and_time(date_value, time_value, combined)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue();
-   }
-
-   return XPathValue(XPathValueType::DateTime, serialise_date_time_components(combined));
-}
-
-XPathValue XPathFunctionLibrary::function_year_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components) or !components.has_date) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.year);
-}
-
-XPathValue XPathFunctionLibrary::function_month_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components) or !components.has_date) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.month);
-}
-
-XPathValue XPathFunctionLibrary::function_day_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components) or !components.has_date) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.day);
-}
-
-XPathValue XPathFunctionLibrary::function_hours_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components) or !components.has_time) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.hour);
-}
-
-XPathValue XPathFunctionLibrary::function_minutes_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components) or !components.has_time) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.minute);
-}
-
-XPathValue XPathFunctionLibrary::function_seconds_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components) or !components.has_time) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue(components.second);
-}
-
-XPathValue XPathFunctionLibrary::function_timezone_from_date_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_time_components(Args[0].to_string(), components)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue();
-   }
-
-   if (!components.has_timezone) return XPathValue();
-
-   return XPathValue(format_timezone_duration(components.timezone_offset_minutes));
-}
-
-XPathValue XPathFunctionLibrary::function_year_from_date(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_value(Args[0].to_string(), components) or !components.has_date) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.year);
-}
-
-XPathValue XPathFunctionLibrary::function_month_from_date(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_value(Args[0].to_string(), components) or !components.has_date) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.month);
-}
-
-XPathValue XPathFunctionLibrary::function_day_from_date(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_value(Args[0].to_string(), components) or !components.has_date) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.day);
-}
-
-XPathValue XPathFunctionLibrary::function_timezone_from_date(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_date_value(Args[0].to_string(), components)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue();
-   }
-
-   if (!components.has_timezone) return XPathValue();
-
-   return XPathValue(format_timezone_duration(components.timezone_offset_minutes));
-}
-
-XPathValue XPathFunctionLibrary::function_hours_from_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_time_value(Args[0].to_string(), components) or !components.has_time) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.hour);
-}
-
-XPathValue XPathFunctionLibrary::function_minutes_from_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_time_value(Args[0].to_string(), components) or !components.has_time) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue((double)components.minute);
-}
-
-XPathValue XPathFunctionLibrary::function_seconds_from_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_time_value(Args[0].to_string(), components) or !components.has_time) {
-      return XPathValue(std::numeric_limits<double>::quiet_NaN());
-   }
-
-   return XPathValue(components.second);
-}
-
-XPathValue XPathFunctionLibrary::function_timezone_from_time(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   DateTimeComponents components;
-   if (!parse_time_value(Args[0].to_string(), components)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue();
-   }
-
-   if (!components.has_timezone) return XPathValue();
-
-   return XPathValue(format_timezone_duration(components.timezone_offset_minutes));
-}
-
-XPathValue XPathFunctionLibrary::function_adjust_date_time_to_timezone(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   std::string value = Args[0].to_string();
-   DateTimeComponents components;
-   if (!parse_date_time_components(value, components)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue(XPathValueType::DateTime, value);
-   }
-
-   bool remove_timezone = false;
-   bool have_target = false;
-   int target_offset = 0;
-
-   if (Args.size() > 1u) {
-      if (Args[1].is_empty()) {
-         remove_timezone = true;
-      }
-      else {
-         int parsed_offset = 0;
-         if (!parse_timezone_duration(Args[1].to_string(), parsed_offset)) {
-            if (Context.expression_unsupported) *Context.expression_unsupported = true;
-            return XPathValue(XPathValueType::DateTime, value);
-         }
-         target_offset = parsed_offset;
-         have_target = true;
-      }
-   }
-
-   if (remove_timezone) {
-      components.has_timezone = false;
-      components.timezone_offset_minutes = 0;
-      components.timezone_is_utc = false;
-      return XPathValue(XPathValueType::DateTime, serialise_date_time_components(components));
-   }
-
-   if (!have_target) {
-      target_offset = 0;
-      have_target = true;
-   }
-
-   std::chrono::sys_time<std::chrono::microseconds> utc_time;
-   if (!components_to_utc_time(components, 0, utc_time)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue(XPathValueType::DateTime, value);
-   }
-
-   DateTimeComponents adjusted = components_from_utc_time(utc_time, target_offset, true, true, true);
-   return XPathValue(XPathValueType::DateTime, serialise_date_time_components(adjusted));
-}
-
-XPathValue XPathFunctionLibrary::function_adjust_date_to_timezone(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   std::string value = Args[0].to_string();
-   DateTimeComponents components;
-   if (!parse_date_value(value, components)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue(XPathValueType::Date, value);
-   }
-
-   bool remove_timezone = false;
-   bool have_target = false;
-   int target_offset = 0;
-
-   if (Args.size() > 1u) {
-      if (Args[1].is_empty()) {
-         remove_timezone = true;
-      }
-      else {
-         int parsed_offset = 0;
-         if (!parse_timezone_duration(Args[1].to_string(), parsed_offset)) {
-            if (Context.expression_unsupported) *Context.expression_unsupported = true;
-            return XPathValue(XPathValueType::Date, value);
-         }
-         target_offset = parsed_offset;
-         have_target = true;
-      }
-   }
-
-   if (remove_timezone) {
-      components.has_timezone = false;
-      components.timezone_offset_minutes = 0;
-      components.timezone_is_utc = false;
-      return XPathValue(XPathValueType::Date, serialise_date_only(components, false));
-   }
-
-   if (!have_target) {
-      target_offset = 0;
-      have_target = true;
-   }
-
-   std::chrono::sys_time<std::chrono::microseconds> utc_time;
-   if (!components_to_utc_time(components, 0, utc_time)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue(XPathValueType::Date, value);
-   }
-
-   DateTimeComponents adjusted = components_from_utc_time(utc_time, target_offset, true, true, false);
-   return XPathValue(XPathValueType::Date, serialise_date_only(adjusted, true));
-}
-
-XPathValue XPathFunctionLibrary::function_adjust_time_to_timezone(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   if (Args.empty() or Args[0].is_empty()) return XPathValue();
-
-   std::string value = Args[0].to_string();
-   DateTimeComponents components;
-   if (!parse_time_value(value, components)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue(XPathValueType::Time, value);
-   }
-
-   bool remove_timezone = false;
-   bool have_target = false;
-   int target_offset = 0;
-
-   if (Args.size() > 1u) {
-      if (Args[1].is_empty()) {
-         remove_timezone = true;
-      }
-      else {
-         int parsed_offset = 0;
-         if (!parse_timezone_duration(Args[1].to_string(), parsed_offset)) {
-            if (Context.expression_unsupported) *Context.expression_unsupported = true;
-            return XPathValue(XPathValueType::Time, value);
-         }
-         target_offset = parsed_offset;
-         have_target = true;
-      }
-   }
-
-   if (remove_timezone) {
-      components.has_timezone = false;
-      components.timezone_offset_minutes = 0;
-      components.timezone_is_utc = false;
-      return XPathValue(XPathValueType::Time, serialise_time_only(components, false));
-   }
-
-   if (!have_target) {
-      target_offset = 0;
-      have_target = true;
-   }
-
-   std::chrono::sys_time<std::chrono::microseconds> utc_time;
-   if (!components_to_utc_time(components, 0, utc_time)) {
-      if (Context.expression_unsupported) *Context.expression_unsupported = true;
-      return XPathValue(XPathValueType::Time, value);
-   }
-
-   DateTimeComponents adjusted = components_from_utc_time(utc_time, target_offset, true, false, true);
-   return XPathValue(XPathValueType::Time, serialise_time_only(adjusted, true));
-}
-
-XPathValue XPathFunctionLibrary::function_implicit_timezone(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   return XPathValue(std::string("PT0S"));
-}
-
-XPathValue XPathFunctionLibrary::function_years_from_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.years.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_months_from_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.months.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_days_from_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.days.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_hours_from_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.hours.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_minutes_from_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.minutes.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_seconds_from_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = components.seconds.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_years_from_year_month_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, true, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.years.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_months_from_year_month_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, true, false);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.months.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_days_from_day_time_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, true);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.days.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_hours_from_day_time_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, true);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.hours.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_minutes_from_day_time_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, true);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = (double)components.minutes.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
-
-XPathValue XPathFunctionLibrary::function_seconds_from_day_time_duration(const std::vector<XPathValue> &Args,
-   const XPathContext &Context)
-{
-   DurationComponents components;
-   DurationParseStatus status = prepare_duration_components(Args, components, false, true);
-
-   if (status IS DurationParseStatus::Empty) return XPathValue();
-   if (status IS DurationParseStatus::Error) return XPathValue(std::numeric_limits<double>::quiet_NaN());
-
-   double value = components.seconds.count();
-   if (components.negative) value = -value;
-   return XPathValue(value);
-}
+#include "functions/func_nodeset.cpp"
+#include "functions/func_strings.cpp"
+#include "functions/func_diagnostics.cpp"
+#include "functions/func_booleans.cpp"
+#include "functions/func_sequences.cpp"
+#include "functions/func_numbers.cpp"
+#include "functions/func_datetimes.cpp"
